+++ native/tests/ae-audio-render.jsx
@@ -25,12 +25,6 @@
         panelFile.open("r");
         var source = panelFile.read();
         panelFile.close();
-        var readingsFile = new File(root + "/native/panel/IslandChatterMandarinReadings.jsxinc");
-        readingsFile.encoding = "UTF-8";
-        readingsFile.open("r");
-        var readingsSource = readingsFile.read();
-        readingsFile.close();
-        eval(readingsSource);
         var bodyStart = source.indexOf("{", source.indexOf("function islandChatterNativePanel("));
         var bodyEnd = source.indexOf("var panel = buildUI(thisObj);");
         eval(source.substring(bodyStart + 1, bodyEnd));